Move type related code to own subdir
