+++ src/type/object.c
@@ -21,7 +21,7 @@
  ******************************************************************************/
 
 
-#include "../include/argent.h"
+#include "../../include/argent.h"
 
 #include <dlfcn.h>
 #include <stdio.h>
