+++ src/type/primitives.c
@@ -21,7 +21,7 @@
  ******************************************************************************/
 
 
-#include "../include/argent.h"
+#include "../../include/argent.h"
 
 #include <errno.h>
 #include <inttypes.h>
