+++ src/type/string.c
@@ -21,7 +21,7 @@
  ******************************************************************************/
 
 
-#include "../include/argent.h"
+#include "../../include/argent.h"
 
 #include <ctype.h>
 #include <regex.h>
