+++ src/type/string.h
