+++ src/type/value.c
@@ -21,7 +21,7 @@
  ******************************************************************************/
 
 
-#include "../include/argent.h"
+#include "../../include/argent.h"
